+++ frontend/src/index.js
@@ -8,7 +8,6 @@ import InstructorView from './views/instructor_view';
 import { AnalysisView } from './views/analysis_view';
 import { DocumentAnalysisView } from "./views/document_analysis_view";
 import ReadingView from './views/reading_view';
-import OverviewView from './views/overview_view';
 
 window.app_modules = {
     React,  // Make React accessible from the base template
@@ -19,6 +18,5 @@ window.app_modules = {
     InstructorView,
     AnalysisView,
     ReadingView,
-    OverviewView,
     DocumentAnalysisView,
 };
